requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.19
PyMuPDF==1.24.9
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.8.0
altair==4.1.0